handled the log status button in feed
