+++ client/src/routes/AppRouter.js
@@ -3,7 +3,6 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 // Pages
 import HomePage from './../pages/HomePage';
-import LandingPage from './../pages/LandingPage';
 import LoginPage from './../pages/LoginPage';
 import SignUpPage from './../pages/SignUpPage';
 import NotFoundPage from './../pages/NotFoundPage';
@@ -14,10 +13,9 @@ class AppRouter extends React.Component {
       <Router>
         <div>
           <Switch>
-            <Route path="/" component={LandingPage} exact={true} />
+            <Route path="/" component={HomePage} exact={true} />
             <Route path="/login" component={LoginPage} />
             <Route path="/signup" component={SignUpPage} />
-            <Route path="/home" component={HomePage} />
             <Route component={NotFoundPage} />
           </Switch>
         </div>
